+++ src/core/compare/primitives.ts
@@ -29,9 +29,9 @@ export function string(x: string, y: string): number {
 	return x.localeCompare(y);
 }
 
-const patterns = Object.freeze({
+const patterns = {
 	'date:complete': /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z)/,
 	'date:time': /\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z)/,
 	date: /\d{4}-[01]\d-[0-3]\d/,
 	time: /[0-2]\d:[0-5]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z)/,
-});
+} as const;