+++ src/std/ntv/object.ts
@@ -1,15 +1,11 @@
 import type { Falsy, IndexSignature } from '../../typings/aliases.js';
-import type { WhenFunction } from '../../typings/comparators.js';
-import type { AnyFunction, Entries } from '../../typings/helpers.js';
+import type { AnyFunction, Entries, Freeze } from '../../typings/helpers.js';
 
 export function entries<T extends object>(o: T) {
 	return Object.entries(o) as Entries<T>;
 }
 
-type DeepFreeze<T> = {
-	readonly [P in keyof T]: WhenFunction<T[P], T[P], DeepFreeze<T[P]>>;
-};
-export function freeze<T extends object>(o: T): DeepFreeze<T> {
+export function freeze<T extends object>(o: T): Freeze<T> {
 	for (const key of Object.getOwnPropertyNames(o)) {
 		const value = o[key as keyof typeof o];
 		if (value == null || typeof value !== 'object') continue;
@@ -18,13 +14,15 @@ export function freeze<T extends object>(o: T): DeepFreeze<T> {
 	return Object.freeze(o);
 }
 
-/** Iterate over objects while retaining its structure */
+/**
+ * Iterate over the key-value pair of an object, returns a new object using the pairs returned from the callback function. If callback is omitted, the default behaviour will create a deep copy of the original object.
+ */
 export function iterate<T extends object>(
 	object: T,
 	callback: AnyFunction<
 		[entry: Entries<T>[number], index: number],
 		void | Falsy | [IndexSignature, any]
-	>
+	> = ([k, v]) => [k, v && typeof v === 'object' ? iterate(v) : v]
 ) {
 	const pairs = entries(object);
 	const memo: typeof pairs = [];
@@ -33,7 +31,7 @@ export function iterate<T extends object>(
 		if (!res || res.length !== 2) continue;
 		memo.push(res as typeof memo[number]);
 	}
-	return Object.fromEntries(memo);
+	return Object.fromEntries(memo) as T;
 }
 
 export function keys<T extends object>(o: T) {