+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # mauss changelog
 
+## Unreleased
+
+- ([#180](https://github.com/alchemauss/mauss/pull/180)) add default parameter to `iterate` callback for object deep copy
+
+## 0.4.4 - 2022/11/29
+
+- ([#182](https://github.com/alchemauss/mauss/pull/182)) add `Freeze` typings
+- ([#179](https://github.com/alchemauss/mauss/pull/179)) fix `clipboard.copy` handler not allowing non-promise callbacks
+
 ## 0.4.3 - 2022/11/25
 
 - ([#177](https://github.com/alchemauss/mauss/pull/177)) fix `Entries` possibly returning `undefined`